requirements.txt
# Data visualization
matplotlib
seaborn
pandas
plotly

# Database and ETL
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
PyYAML>=6.0
tqdm>=4.65.0
tabulate>=0.9.0

# Optional: For Parquet export
pyarrow>=12.0.0

# Optional: For faster data processing
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0

# Jupyter and Interactive Development
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0

# Additional ML Visualization
shap>=0.45.0  # Feature importance and model interpretation

# Interactive Dashboard
streamlit>=1.31.0  # Main dashboard framework
streamlit-plotly-events>=0.0.6  # Interactive Plotly charts in Streamlit
pdf2image>=1.16.3  # Convert PDF track maps to images
Pillow>=10.2.0  # Image processing for track visualizations